+++ packages/wonder-blocks-modal/src/components/modal-dialog.tsx
@@ -2,6 +2,7 @@ import * as React from "react";
 import {View} from "@khanacademy/wonder-blocks-core";
 import type {StyleType} from "@khanacademy/wonder-blocks-core";
 import {StyleSheet} from "aphrodite";
+import {semanticColor} from "@khanacademy/wonder-blocks-tokens";
 import theme from "../theme";
 
 type Props = {
@@ -94,14 +95,17 @@ const small = "@media (max-width: 767px)" as any;
 
 const styles = StyleSheet.create({
     wrapper: {
-        display: "flex",
+        // Allows propagating the text color to all the children.
+        color: semanticColor.core.foreground.neutral.strong,
         flexDirection: "row",
         alignItems: "stretch",
         width: "100%",
         height: "100%",
         position: "relative",
+        boxShadow: theme.dialog.shadow.default,
+        borderRadius: theme.root.border.radius,
         [small]: {
-            padding: theme.dialog.spacing.padding,
+            padding: theme.dialog.layout.padding,
             flexDirection: "column",
         },
     },