+++ packages/wonder-blocks-tokens/src/theme/semantic/semantic-color.ts
@@ -1,5 +1,5 @@
-// This use is valid as this is the only place in the codebase where we should
-// import/reference primitive colors.
+// This import is valid as this is the only place in the codebase where we
+// should import/reference primitive colors.
 /* eslint-disable import/no-deprecated */
 import {color} from "../../tokens/color";
 // NOTE: This import is only to provide backwards compatibility with some