+++ packages/wonder-blocks-breadcrumbs/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @khanacademy/wonder-blocks-breadcrumbs
 
+## 3.1.22
+
+### Patch Changes
+
+- Updated dependencies [e457d8c]
+    - @khanacademy/wonder-blocks-tokens@11.2.0
+
+## 3.1.21
+
+### Patch Changes
+
+- Updated dependencies [f26858a]
+    - @khanacademy/wonder-blocks-tokens@11.1.1
+
 ## 3.1.20
 
 ### Patch Changes