+++ packages/wonder-blocks-clickable/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @khanacademy/wonder-blocks-clickable
 
+## 7.1.11
+
+### Patch Changes
+
+- Updated dependencies [e457d8c]
+    - @khanacademy/wonder-blocks-tokens@11.2.0
+
+## 7.1.10
+
+### Patch Changes
+
+- Updated dependencies [f26858a]
+    - @khanacademy/wonder-blocks-tokens@11.1.1
+
 ## 7.1.9
 
 ### Patch Changes