+++ packages/wonder-blocks-dropdown/CHANGELOG.md
@@ -1,5 +1,52 @@
 # @khanacademy/wonder-blocks-dropdown
 
+## 10.1.13
+
+### Patch Changes
+
+- Updated dependencies [e457d8c]
+- Updated dependencies [e457d8c]
+    - @khanacademy/wonder-blocks-modal@8.1.0
+    - @khanacademy/wonder-blocks-tokens@11.2.0
+    - @khanacademy/wonder-blocks-cell@4.1.25
+    - @khanacademy/wonder-blocks-clickable@7.1.11
+    - @khanacademy/wonder-blocks-icon@5.2.5
+    - @khanacademy/wonder-blocks-icon-button@10.3.6
+    - @khanacademy/wonder-blocks-layout@3.1.22
+    - @khanacademy/wonder-blocks-pill@3.1.25
+    - @khanacademy/wonder-blocks-search-field@5.1.28
+    - @khanacademy/wonder-blocks-typography@4.2.7
+
+## 10.1.12
+
+### Patch Changes
+
+- Updated dependencies [f26858a]
+    - @khanacademy/wonder-blocks-tokens@11.1.1
+    - @khanacademy/wonder-blocks-cell@4.1.24
+    - @khanacademy/wonder-blocks-clickable@7.1.10
+    - @khanacademy/wonder-blocks-icon@5.2.4
+    - @khanacademy/wonder-blocks-icon-button@10.3.5
+    - @khanacademy/wonder-blocks-layout@3.1.21
+    - @khanacademy/wonder-blocks-modal@8.0.1
+    - @khanacademy/wonder-blocks-pill@3.1.24
+    - @khanacademy/wonder-blocks-search-field@5.1.27
+    - @khanacademy/wonder-blocks-typography@4.2.6
+
+## 10.1.11
+
+### Patch Changes
+
+- Updated dependencies [705ee01]
+- Updated dependencies [9ef528a]
+- Updated dependencies [038f9a9]
+    - @khanacademy/wonder-blocks-typography@4.2.5
+    - @khanacademy/wonder-blocks-modal@8.0.0
+    - @khanacademy/wonder-blocks-cell@4.1.23
+    - @khanacademy/wonder-blocks-icon-button@10.3.4
+    - @khanacademy/wonder-blocks-pill@3.1.23
+    - @khanacademy/wonder-blocks-search-field@5.1.26
+
 ## 10.1.10
 
 ### Patch Changes