+++ packages/wonder-blocks-icon/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @khanacademy/wonder-blocks-icon
 
+## 5.2.5
+
+### Patch Changes
+
+- Updated dependencies [e457d8c]
+    - @khanacademy/wonder-blocks-tokens@11.2.0
+
+## 5.2.4
+
+### Patch Changes
+
+- Updated dependencies [f26858a]
+    - @khanacademy/wonder-blocks-tokens@11.1.1
+
 ## 5.2.3
 
 ### Patch Changes