+++ packages/wonder-blocks-link/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @khanacademy/wonder-blocks-link
 
+## 9.1.11
+
+### Patch Changes
+
+- Updated dependencies [e457d8c]
+    - @khanacademy/wonder-blocks-tokens@11.2.0
+    - @khanacademy/wonder-blocks-clickable@7.1.11
+    - @khanacademy/wonder-blocks-icon@5.2.5
+
+## 9.1.10
+
+### Patch Changes
+
+- Updated dependencies [f26858a]
+    - @khanacademy/wonder-blocks-tokens@11.1.1
+    - @khanacademy/wonder-blocks-clickable@7.1.10
+    - @khanacademy/wonder-blocks-icon@5.2.4
+
 ## 9.1.9
 
 ### Patch Changes