+++ packages/wonder-blocks-search-field/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @khanacademy/wonder-blocks-search-field
 
+## 5.1.28
+
+### Patch Changes
+
+- Updated dependencies [e457d8c]
+    - @khanacademy/wonder-blocks-tokens@11.2.0
+    - @khanacademy/wonder-blocks-form@7.1.25
+    - @khanacademy/wonder-blocks-icon@5.2.5
+    - @khanacademy/wonder-blocks-icon-button@10.3.6
+    - @khanacademy/wonder-blocks-typography@4.2.7
+
+## 5.1.27
+
+### Patch Changes
+
+- Updated dependencies [f26858a]
+    - @khanacademy/wonder-blocks-tokens@11.1.1
+    - @khanacademy/wonder-blocks-form@7.1.24
+    - @khanacademy/wonder-blocks-icon@5.2.4
+    - @khanacademy/wonder-blocks-icon-button@10.3.5
+    - @khanacademy/wonder-blocks-typography@4.2.6
+
+## 5.1.26
+
+### Patch Changes
+
+- Updated dependencies [705ee01]
+    - @khanacademy/wonder-blocks-typography@4.2.5
+    - @khanacademy/wonder-blocks-form@7.1.23
+    - @khanacademy/wonder-blocks-icon-button@10.3.4
+
 ## 5.1.25
 
 ### Patch Changes