+++ packages/wonder-blocks-styles/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @khanacademy/wonder-blocks-styles
 
+## 0.2.18
+
+### Patch Changes
+
+- Updated dependencies [e457d8c]
+    - @khanacademy/wonder-blocks-tokens@11.2.0
+
+## 0.2.17
+
+### Patch Changes
+
+- Updated dependencies [f26858a]
+    - @khanacademy/wonder-blocks-tokens@11.1.1
+
 ## 0.2.16
 
 ### Patch Changes