+++ packages/wonder-blocks-switch/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @khanacademy/wonder-blocks-switch
 
+## 3.3.4
+
+### Patch Changes
+
+- Updated dependencies [e457d8c]
+    - @khanacademy/wonder-blocks-tokens@11.2.0
+    - @khanacademy/wonder-blocks-icon@5.2.5
+    - @khanacademy/wonder-blocks-styles@0.2.18
+
+## 3.3.3
+
+### Patch Changes
+
+- Updated dependencies [f26858a]
+    - @khanacademy/wonder-blocks-tokens@11.1.1
+    - @khanacademy/wonder-blocks-icon@5.2.4
+    - @khanacademy/wonder-blocks-styles@0.2.17
+
 ## 3.3.2
 
 ### Patch Changes