+++ packages/wonder-blocks-tabs/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @khanacademy/wonder-blocks-tabs
 
+## 0.3.14
+
+### Patch Changes
+
+- Updated dependencies [e457d8c]
+    - @khanacademy/wonder-blocks-tokens@11.2.0
+    - @khanacademy/wonder-blocks-typography@4.2.7
+
+## 0.3.13
+
+### Patch Changes
+
+- 65e0b82: Adds `flexShrink: 0` to tablist wrapper styles so that it does not vanish
+
+## 0.3.12
+
+### Patch Changes
+
+- Updated dependencies [f26858a]
+    - @khanacademy/wonder-blocks-tokens@11.1.1
+    - @khanacademy/wonder-blocks-typography@4.2.6
+
+## 0.3.11
+
+### Patch Changes
+
+- Updated dependencies [705ee01]
+    - @khanacademy/wonder-blocks-typography@4.2.5
+
 ## 0.3.10
 
 ### Patch Changes