+++ packages/wonder-blocks-toolbar/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @khanacademy/wonder-blocks-toolbar
 
+## 5.1.24
+
+### Patch Changes
+
+- Updated dependencies [e457d8c]
+    - @khanacademy/wonder-blocks-tokens@11.2.0
+    - @khanacademy/wonder-blocks-typography@4.2.7
+
+## 5.1.23
+
+### Patch Changes
+
+- Updated dependencies [f26858a]
+    - @khanacademy/wonder-blocks-tokens@11.1.1
+    - @khanacademy/wonder-blocks-typography@4.2.6
+
+## 5.1.22
+
+### Patch Changes
+
+- Updated dependencies [705ee01]
+    - @khanacademy/wonder-blocks-typography@4.2.5
+
 ## 5.1.21
 
 ### Patch Changes